requirements.txt
click
google-generativeai
httpx
beautifulsoup4
lxml
python-dotenv
ddgs
youtube-transcript-api
flask
